1.5

+++ jd-manager/jd-manager-interface/src/main/java/com/zhou/jdshop/service/ProductService.java
@@ -11,7 +11,7 @@ public interface ProductService {
 
     List<TbProductCustom> listProducts();
 
-    int saveProduct(TbProduct product);
+    Long saveProduct(TbProduct product);
 
     int updateProduct(List<Long> pids,Integer pflag);
